Multiplicity Keyword Change, New Function for Multiplicity Time Evolution and other minor fixes
The multiplicity keyword is changed from Fraction or Frequency to MF or CF in all instances. The Multiplicity Time Evolution is changed from different functions for CF/MF to one function for both (this will be implemented for other functions later), and a few more minor fixes in other parts of the code.